+++ huey/tests/__init__.py
@@ -1,4 +1,5 @@
 from huey.tests.test_api import *
+from huey.tests.test_asyncio import *
 from huey.tests.test_consumer import *
 from huey.tests.test_crontab import *
 from huey.tests.test_django_tasks import *